Implement deferred imports for 60x faster openadmet.models.registries load
Baseline: import openadmet.models.registries = 6.702s
After: import openadmet.models.registries = 0.111s (60x faster)

Phase 1 - Split model_base.py:
- Create architecture/lightning_model_base.py isolating all torch/lightning imports
- Strip model_base.py of torch/lightning/joblib top-level imports
- Add PEP 562 module __getattr__ for lazy LightningModelBase re-export
- Defer joblib inside save()/load() method bodies
- Result: architecture/model_base.py 1.652s -> 0.070s

Phase 2 - Deferred estimator class imports:
- Replace mod_class: ClassVar[type] = SomeClass with _get_estimator_class()
  classmethod in all concrete architecture modules (xgboost, catboost, lgbm,
  rf, svm, tabpfn, dummy)
- Remove all top-level 3rd-party imports from these modules
- Move _METRIC_TO_LOSS dict initialization inside chemprop build()
- Result: each arch module 2-3s -> ~0.1s

Phase 3 - Deferred imports in features/split/trainer/eval:
- feature_base.py: move molfeat/torch imports to TYPE_CHECKING block
- features/chemprop.py: remove self-import bug; defer all chemprop/torch/sklearn
  imports inside featurize() and _vendor_build_dataloader()
- split/scaffold.py: defer splito and sklearn.model_selection inside split()
- split/cluster.py: defer useful_rdkit_utils, datamol, molfeat, KMeans inside
  split(); remove unused GroupShuffleSplit import
- trainer/lightning.py: defer torch and lightning imports inside build()/train()
- eval/regression.py: defer wandb, scipy.stats, sklearn.metrics, seaborn inside
  their respective usage methods; convert _metrics class var to _base_metrics()
  classmethod; fix cross_validation.py to not import removed module-level names
- eval/eval_base.py: defer scipy.stats.bootstrap inside stat_and_bootstrap()
- Result: registries 6.702s -> 3.548s

Phase 4 - Lazy registry loading:
- Create _registry_loader.py with idempotent load_group()/load_all() functions
  using importlib.import_module; zero heavy imports at module level
- Rewrite registries.py to only import base registry objects and expose load_all()
- Add load_group() call to each get_*_class() function for on-demand loading
- Update anvil/specification.py and anvil/workflow_base.py to import load_all
  instead of wildcard-importing registries
- Result: import openadmet.models.registries 6.702s -> 0.111s (60x faster)

Before/after summary:
  import openadmet.models.registries: 6.702s -> 0.111s
  architecture/model_base.py:         1.652s -> 0.070s
  architecture/xgboost.py:            2.123s -> 0.099s
  architecture/chemprop.py:           3.083s -> 0.101s
  split/cluster.py:                   3.524s -> 0.331s
  split/scaffold.py:                  1.476s -> 0.330s
  trainer/lightning.py:               1.653s -> 0.069s
  eval/regression.py:                 1.582s -> 0.326s
  registries + load_all():            N/A    -> 3.727s (same real cost, deferred)
