+++ openadmet/models/architecture/svm.py
@@ -4,7 +4,6 @@ from typing import ClassVar
 
 import numpy as np
 from loguru import logger
-from sklearn.svm import SVC, SVR
 
 from openadmet.models.architecture.model_base import PickleableModelBase, models
 
@@ -14,12 +13,16 @@ class SVMModelBase(PickleableModelBase):
 
     # Meta parameters for this class
     type: ClassVar[str]
-    mod_class: ClassVar[type]
+
+    @classmethod
+    def _get_estimator_class(cls) -> type:
+        """Return the SVM estimator class (deferred import)."""
+        raise NotImplementedError
 
     def build(self):
         """Prepare the model."""
         if not self.estimator:
-            self.estimator = self.mod_class(**self.model_dump())
+            self.estimator = self._get_estimator_class()(**self.model_dump())
         else:
             logger.warning("Model already exists, skipping build")
 
@@ -80,7 +83,12 @@ class SVMRegressorModel(SVMModelBase):
 
     # Meta parameters for this class
     type: ClassVar[str] = "SVMRegressorModel"
-    mod_class: ClassVar[type] = SVR
+
+    @classmethod
+    def _get_estimator_class(cls) -> type:
+        from sklearn.svm import SVR
+
+        return SVR
 
     # SVR parameters
     kernel: str = "rbf"
@@ -115,7 +123,12 @@ class SVMClassifierModel(SVMModelBase):
 
     # Meta parameters for this class
     type: ClassVar[str] = "SVMClassifierModel"
-    mod_class: ClassVar[type] = SVC
+
+    @classmethod
+    def _get_estimator_class(cls) -> type:
+        from sklearn.svm import SVC
+
+        return SVC
 
     # SVC parameters
     C: float = 1.0
