+++ openadmet/models/trainer/lightning.py
@@ -3,10 +3,6 @@
 from pathlib import Path  # it is used in the main therefore i do not remove it
 from typing import Any
 
-import torch
-from lightning import pytorch as pl
-from lightning.pytorch.callbacks import EarlyStopping, ModelCheckpoint
-from lightning.pytorch.loggers import CSVLogger, WandbLogger
 from loguru import logger
 from openadmet.models.drivers import DriverType
 from openadmet.models.trainer.trainer_base import TrainerBase, trainers
@@ -106,6 +102,10 @@ class LightningTrainer(TrainerBase):
         # Initialize logging container
         self._logger = []
 
+        from lightning import pytorch as pl
+        from lightning.pytorch.callbacks import EarlyStopping, ModelCheckpoint
+        from lightning.pytorch.loggers import CSVLogger, WandbLogger
+
         # Initialize the callbacks dict
         self._callbacks = {}
 
@@ -190,6 +190,8 @@ class LightningTrainer(TrainerBase):
         """
         # Indicate that the model is being trained
         logger.debug(f"Training model {self.model.estimator}")
+
+        import torch
 
         # Fit model
         self._trainer.fit(self.model.estimator, train_dataloader, val_dataloader)
